+++ src/shared/meson.build
@@ -105,6 +105,8 @@ shared_sources = files(
         'dns-domain.h',
         'dropin.c',
         'dropin.h',
+        'efi-api.c',
+        'efi-api.h',
         'efi-loader.c',
         'efi-loader.h',
         'elf-util.c',
